[#42296389] Initial manifests for rbenv-ruby-1.9.2
Installs Ruby to a prefix directory. The parent of which is owned by the
rbenv package, which is listed as a dependency.

Package name contains the entire version string so that we can have multiple
unique versions (e.g. 1.9.2-p290, 1.9.2-p320, 1.9.3-p392) in the same repo.

Reference the patch which we've been using to-date.